fwk: add fn key combinations for row B end keys
Maps fn+Z to the ISO 102nd key, and fn+/ to the extra key
left of right shift on Brazilian and Japanese layouts.
This allows full use of ISO and Brazilian layouts on
ANSI, ISO, or JIS keyboard hardware. (Alternately,
users can remap these keys in the OS to whatever function
they prefer.)

Signed-off-by: Jules Bertholet <julesbertholet@quoi.xyz>