fwk: report fn key and lock states via HID
Fn key state is reported as "System Function Shift",
FnLock *key* (Fn+Esc) state as "System Function Shift Lock",
and Fn lock state is reported as "System Function Shift
Lock Indicator". This allows the OS to detect when the Fn
key is pressed, enabling the user to configure custom keybinds
involving it; it also allows the OS to report the state of the
Fn lock feature to the user.

Signed-off-by: Jules Bertholet <julesbertholet@quoi.xyz>